Moved common truncation logic to method in base class
